[ init ] A few libraries, a few examples
Data.Infinities and Data.UniqueSortedList are quite clearly heavily
influenced by McBride's "How to keep your neighbours in order".
Here insertion may fail as we are guaranteeing that no two equal
values (for the notion of equality of the strict total order) are
stored in the list. We do not want a flag to have multiple meanings...