[ refactor ] Data.Table.show as a display method
Rather than hand-writing the function displaying the content in a neatly
formatted table, we can abstract the notion of pretty-printing a Table
and use that.

This allows us to shrink the WordCount example a bit by removing the
code which has not much to do with the logic of the app.