[ fix ] Fields are not functions anymore
In order for more things to be considered equal by the typechecker,
it is necessary to move from Fields as functions to Fields as
iterated product of Sets.

All the previous work has now been ported to the new definition and
the Data/Record/Examples are now cast-free!

Related thread:
http://code.google.com/p/agda/issues/detail?id=1526#c1