[ fix ] using record for Command to have eta rules
The ParsedCommmand datatype was quite hard to inhabit because of the
akward return types in the index. Now that we have records and thus
eta rules, it's all a lot smoother.

I'm also playing the usual tricks when it comes to convincing the
termination checker in the various parseXXX functions (that is state
the computations you are potentially interested in and then only use
the ones you really need).