[ new ] Smart constructors for Records
This time we have smart constructors which are *always* successful.
This means that there is no need for a {pr : T _} argument anymore
but, more importantly, that inference works a lot better as
demonstrated by the removal of redundant information in Sum and
Wordcount.

It is still possible to state that a given record ought to have a
given type by adding a top level annotation.